# src/main/resources/api-contract/library-api-contract.yaml
openapi: 3.0.0
servers:
  - description: API Auto Mocking
    url: https://virtserver.swaggerhub.com/bharatiyas/library-api/1.0.0
info:
  description: |
    This is API specification for Library APIs.
  version: "1.0.0"
  title: Library API
  contact:
    email: sanjayfromgomi@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /authors:
    post:
      tags:
        - authors
      summary: Add a new author to the library
      operationId: addAuthor
      responses:
        '201':
          description: Author created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: Bad Request
        '409':
          description: Author already exists
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthor'
  /authors/{authorId}:
    get:
      tags:
        - authors
      summary: Find author by ID
      description: Returns a single author wth Id supplied
      operationId: getAuthorsById
      parameters:
        - name: authorId
          in: path
          description: ID of author to return
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Author found by the supplied ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Author not found
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error

    put:
      tags:
        - authors
      summary: Update an existing author. It can be used to update email and phone number of the author
      operationId: updateAuthor
      parameters:
        - name: authorId
          in: path
          description: AuthorId to update
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Author updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAuthor'
        '400':
          description: Bad Request
        '404':
          description: Author not found
        '406':
          description: Method not allowed
        '409':
          description: Author already exists
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'

    delete:
      tags:
        - authors
      summary: Deletes an Author
      operationId: deleteAuthor
      parameters:
        - name: authorId
          in: path
          description: AuthorId to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: Author deleted successfully
        '404':
          description: Author not found
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error
  /authors/search:
    get:
      tags:
        - authors
      summary: Search authors
      description: Search authors based on firstName and lastName
      operationId: searchAuthors
      parameters:
        - name: firstName
          in: query
          description: First name of the author to be searched
          required: true
          explode: true
          schema:
            type: string
            example: "Herbet"
        - name: lastName
          in: query
          description: Last name of the author to be searched
          required: true
          explode: true
          schema:
            type: string
            example: "Shield"
      responses:
        '200':
          description: Authors list matching the first and last name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        '400':
          description: Bad Request. No search criteria supplied
        '404':
          description: No Authors found
        '406':
          description: Method not allowed
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    CreateAuthor:
      type: object
      properties:
        firstName:
          type: string
          example: "TestauthorFn"
        lastName:
          type: string
          example: "TestauthorLn"
        dateOfBirth:
          type: string
          format: date
          example: "2000-01-20"
        gender:
          $ref: '#/components/schemas/Gender'
    UpdateAuthor:
      type: object
      properties:
        dateOfBirth:
          type: string
          format: date
          example: "2000-01-20"
        gender:
          $ref: '#/components/schemas/Gender'
    Author:
      type: object
      properties:
        authorId:
          type: integer
          example: 8973
        firstName:
          type: string
          example: "TestauthorFn"
        lastName:
          type: string
          example: "TestauthorLn"
        dateOfBirth:
          type: string
          format: date
          example: "2000-01-20"
        gender:
          $ref: '#/components/schemas/Gender'
    Gender:
      enum:
        - Male
        - Female
        - Undisclosed
...